added necessary comments and removed unused files
